chore: deps upgrade & added workspaces to avoid linking packages
