+++ .eslintignore
@@ -1,0 +1,3 @@
+**/build/
+**/node_modules/
+.tmp
