+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -9,7 +9,7 @@ body:
         This issue tracker is for reporting bugs found in `remix-themes` (https://github.com/abereghici/remix-themes).
         If you have a question about how to achieve something and are struggling, please post a question
         inside of `remix-themes` Discussions tab: https://github.com/abereghici/remix-themes/discussions
-        
+
         Before submitting a new bug/issue, please check the links below to see if there is a solution or question posted there already:
          - `remix-themes` Issues tab: https://github.com/abereghici/remix-themes/issues?q=is%3Aissue+is%3Aopen+sort%3Aupdated-desc
          - `remix-themes` closed issues tab: https://github.com/abereghici/remix-themes/issues?q=is%3Aissue+sort%3Aupdated-desc+is%3Aclosed
@@ -20,7 +20,9 @@ body:
     id: description
     attributes:
       label: Describe the bug
-      description: Provide a clear and concise description of the challenge you are running into.
+      description:
+        Provide a clear and concise description of the challenge you are running
+        into.
     validations:
       required: true
   - type: input
@@ -53,7 +55,8 @@ body:
     id: expected
     attributes:
       label: Expected behavior
-      description: Provide a clear and concise description of what you expected to happen.
+      description:
+        Provide a clear and concise description of what you expected to happen.
       placeholder: |
         As a user, I expected ___ behavior but i am seeing ___
     validations:
