+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,8 +1,8 @@
 blank_issues_enabled: false
 contact_links:
-  - name: 🤔 Feature Requests & Questions 
+  - name: 🤔 Feature Requests & Questions
     url: https://github.com/abereghici/remix-themes/discussions
     about: Please ask and answer questions here.
-  - name: 💬 Remix Discord Channel 
+  - name: 💬 Remix Discord Channel
     url: https://rmx.as/discord
     about: Please ask and answer questions here.
