--- demo/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": ["@remix-run/eslint-config", "@remix-run/eslint-config/node"]
-}
