--- demo/app/entry.client.tsx
@@ -1,4 +1,0 @@
-import { RemixBrowser } from "@remix-run/react";
-import { hydrate } from "react-dom";
-
-hydrate(<RemixBrowser />, document);
