+++ packages/remix-themes-app/app/entry.client.tsx
@@ -1,0 +1,4 @@
+import {RemixBrowser} from '@remix-run/react'
+import {hydrateRoot} from 'react-dom/client'
+
+hydrateRoot(document, <RemixBrowser />)
