+++ packages/remix-themes-app/app/entry.server.tsx
@@ -1,0 +1,21 @@
+import type {EntryContext} from '@remix-run/node'
+import {RemixServer} from '@remix-run/react'
+import {renderToString} from 'react-dom/server'
+
+export default function handleRequest(
+  request: Request,
+  responseStatusCode: number,
+  responseHeaders: Headers,
+  remixContext: EntryContext,
+) {
+  let markup = renderToString(
+    <RemixServer context={remixContext} url={request.url} />,
+  )
+
+  responseHeaders.set('Content-Type', 'text/html')
+
+  return new Response('<!DOCTYPE html>' + markup, {
+    status: responseStatusCode,
+    headers: responseHeaders,
+  })
+}
