+++ packages/remix-themes-app/app/sessions.server.tsx
@@ -1,15 +1,16 @@
 import {createCookieSessionStorage} from '@remix-run/node'
 import {createThemeSessionResolver} from 'remix-themes'
+
 export const themeSessionResolver = createThemeSessionResolver(
   createCookieSessionStorage({
     cookie: {
-      name: 'remix-themes',
-      secure: true,
+      name: '__remix-themes',
+      domain: 'localhost',
+      path: '/',
+      httpOnly: true,
       sameSite: 'lax',
       secrets: ['s3cr3t'],
-      path: '/',
-      expires: new Date('2100-08-14'),
-      httpOnly: true,
+      // secure: true,
     },
   }),
 )
