+++ packages/remix-themes-app/public/favicon.ico
