+++ packages/remix-themes-app/remix.config.js
@@ -1,0 +1,4 @@
+/** @type {import('@remix-run/dev').AppConfig} */
+module.exports = {
+  ignoredRouteFiles: ['**/.*'],
+}
