+++ packages/remix-themes/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "remix-themes",
+  "version": "0.0.0-semantically-released",
+  "description": "An abstraction for themes in your Remix_run app.",
+  "keywords": [
+    "remix",
+    "remix.run",
+    "react",
+    "themes",
+    "dark mode",
+    "light mode"
+  ],
+  "homepage": "https://github.com/abereghici/remix-themes#readme",
+  "bugs": {
+    "url": "https://github.com/abereghici/remix-themes/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/abereghici/remix-themes.git"
+  },
+  "license": "MIT",
+  "author": "Alexandru Bereghici",
+  "main": "./build/index.js",
+  "types": "./build/index.d.ts",
+  "files": [
+    "build",
+    "package.json",
+    "README.md"
+  ],
+  "scripts": {
+    "build": "tsc --project tsconfig.json",
+    "typecheck": "tsc --project tsconfig.json --noEmit",
+    "lint": "eslint --ext .ts,.tsx src/",
+    "validate": "npm-run-all --parallel typecheck lint"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@remix-run/react": "latest",
+    "@remix-run/server-runtime": "latest",
+    "@types/react": "^18.0.27",
+    "npm-run-all": "^4.1.5",
+    "react": "^18.2.0",
+    "typescript": "^4.9.5"
+  },
+  "peerDependencies": {
+    "@remix-run/react": "^1.6.0",
+    "@remix-run/server-runtime": "^1.6.0"
+  }
+}
