+++ prettier.config.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line unicorn/prefer-module
 module.exports = {
   arrowParens: 'avoid',
   bracketSpacing: false,
